==> Figure6_B_F/02_hts2_analysis_Fig6.Rmd <==
---
title:   "FGFR2 HCR siKinome Screen - No4 - Part 2 - cellHTS2 Data analysis"
author: "Asaf Shilo/Gianluca Pegoraro"
date: "May 12th 2022"
output: github_document
---
## README
This script is used to analyze siRNA screen data obtained with High-Content Image analysis. In particular, the siRNA library used is the ThermoFisher siKinome. The library contains 3 oligo siRNA sequences per gene against 709 genes that have kinase domains. 
The library was originally received dried in 96 well plates at 0.25 nM concentration. It was resuspended in 50 ul ddH20 (5 uM final), frozen and then thawed to increase oligo siRNA solubility. Column 12 of the 96-well plates was left empty on purpose.  The resuspended libraries (27 plates total), were then trasnferred and compressed into 384-well format to obtain 7 Mother plates (Plate 7 only contains 5 oligos and was left out from the final analysis, hence we will have 704 genes at the end, not 709). For every mother plate, 3 384-LDV daughter plates (13 ul each) were generated at the same concentration. Columns 23 and 24 were originally left empty, and siRNA controls were added at a later stage. All these steps were performed using a PE Janus, which output all the liquid handling operations logs as text files

Imaging assay-ready plates were generated by spotting 150 nl of siRNA oligo at the bottom of each plate using an Echo525. Imaging Assay plates were dried, frozen and then used in reverse transfection experiments. For reverese transfection, plates were thawed, spinned and 20 ul of Optimem + 0.05 ul/well of RNAiMax were added to the plates and incubated for 30'. Then 20 ul of cells were added on top of the siRNA/RNAiMax mix and incubated for 72 hrs. 

Fixed and stained plates were imaged on an CV7000S microscope using a 40X air objectives and Z-stacks with slice interval of 1 micron. Z-stacks were maximally projected on the fly. Images were analyzed in Columbus 2.8.1, and image analysis results were exported as tab delimited .txt files. 

The script reads the Janus logs files, the Columbus image analysis results and generates all the necessary .txt files for the analysis of the siRNA screen results in `cellHTS2`. In addition, it also runs the `cellHTS2` analysis and outputs an html report, .txt results tables, and some custom plots of the z-score distributions.

## Setup
Load packages.
```{r}
library(tidyverse)
library(cellHTS2)
```
Set Knitr options.

```{r, include=FALSE, warning=FALSE}
knitr::opts_chunk$set(
  message = FALSE,
  warning = FALSE
  )
```

### User variables input and settings specification
Choose a Columbus image analysis result variable that needs to be analized in `cellHTS2`. This might change from analysis to analysis.
The chunk below assigns the Columbus variable chosen for analysis and sets the `cellHTS2` input and output directories names accordingly . 


```{r}
well_tbl_2 <- read_tsv("HTS2_Input/Columbus_object_results.txt")
annotation_tbl <- read_tsv("HTS2_Metadata/Annotation.txt")

var_names <- well_tbl_2 %>% select(cell_n:FGFR2_Full_spots_TBP_norm) %>% names()
```

## `cellHTS2` data analysis
```{r importPlateList}
##`cellHTS2` data analysis function

calculate_HTS2 <- function(meas_select, in_dir = NULL, metadata_dir = NULL, out_dir = NULL){
  
  #Read the measurement files in the in_dir directory. 
  obj <- readPlateList(filename = "Platelist.txt",
                             name = meas_select,
                             path = file.path(in_dir, meas_select))
  
  #Configure the cellHTS2 object
  obj <- configure(obj,
                   descripFile = "Description.txt",
                   confFile = "Plateconf.txt",
                   path = metadata_dir)
  
  #Normalize the measurements on a per plate-basis.
  obj_n <- normalizePlates(obj,
                          scale = "additive",
                          log = FALSE,
                          method = "Bscore", #Spatial normalization
                          varianceAdjust = "byPlate")

  #Score the replicates by taking the Z-score of the B-score distribution 
  # calculated in the previous chunk.
  obj_sc <- scoreReplicates(obj_n, 
                            sign = "+", 
                            method = "zscore")
  
  #Summarize the replicates. In this particular case the final Z-score is going 
  # to be the mean of the two replicates Z-scores. 
  obj_sc <-summarizeReplicates(obj_sc, 
                               summary = "mean")
  
  # Annotate the cellHTS2 object with gene names and siRNA id's. 
  obj_sc <- cellHTS2::annotate(obj_sc,
                               geneIDFile = "Annotation.txt",
                               path = metadata_dir)

  #Save the cellHTS2 object as an `.rda` file in the appropriate results 
  #subfolder in case someone wants to inspect it and/or further process it.
  save(obj_sc, file = file.path(out_dir, 
                                meas_select, 
                                "obj_sc.rda"))
  
    
  setSettings(list(plateList = list(reproducibility = list(include = TRUE, 
                                                           map = TRUE),
                                    intensities = list(include = TRUE, map = TRUE)),
                   screenSummary = list(scores = list(range = c(-8, 8), map = TRUE))))
  
  # Write the report in the appropriate output directory
  cellHTS2_report <- writeReport(raw = obj,
                                normalized = obj_n,
                                scored = obj_sc,
                                outdir = file.path(out_dir, 
                                                   meas_select),
                                force = TRUE,
                                mainScriptFile = "02_hts2_analysis_Fig6.Rmd")
  
  #Write the results table in the appropriate output directory                             
  results_tbl <- getTopTable(cellHTSlist = list("raw" = obj,
                                                "normalized" = obj_n,
                                                "scored" = obj_sc),
                            file = file.path(out_dir, 
                                             meas_select, 
                                             "Results_table.txt"))
  
  #Calculate and write the median strongest siRNA per gene (Based on Z-scores)
  results_median_tbl <- results_tbl %>% 
                        filter(!(is.na(GeneSymbol))) %>%
                        group_by(GeneID, GeneSymbol) %>%
                        summarise(median_z_score = median(score)) %>%
                        arrange(desc(median_z_score)) %>%
                        write_tsv(file.path(out_dir, 
                                            meas_select, 
                                            "Results_median_table.txt"))
  #Calculate and write the RSA scores
  rsa_tbl <- rsa(obj_sc, reverse = T) %>%
             filter(!(is.na(GeneID))) %>%
             mutate(Plate = as.integer(Plate)) %>%
             left_join(annotation_tbl, 
                       by = c("Plate", "Well")) %>%
             arrange(RSARank) %>%
             write_tsv(file.path(out_dir, 
                                meas_select, 
                                "Results_RSA_table.txt"))
  rsa_tbl_filter <- rsa_tbl %>%
                    mutate(Flag = Score > 1.5) %>%
                    group_by(GeneID.x) %>%
                    summarise(Flag = sum(Flag) >= 2)
                  
  rsa_filtered_tbl <- rsa_tbl %>%
                      left_join(rsa_tbl_filter, by = "GeneID.x") %>%
                      filter(Flag)

  write_tsv(rsa_filtered_tbl, 
            file.path(out_dir, 
                      meas_select, 
                      "Results_Filtered_RSA_table.txt"))
}
```

```{r}
var_names %>% walk(calculate_HTS2, in_dir = "hts2_input/", metadata_dir = "hts2_metadata/", out_dir = "hts2_output/")
```

Document the information about the analysis session.

```{r sessionInfo, results='markup'}
sessionInfo()
```
